pyproject.toml
[project]
name = "ats-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
